start with adding ckediter plugin
I have used these instruction: https://strapi.io/documentation/developer-docs/latest/guides/registering-a-field-in-admin.html

The problem I am deeling with is the CKEditer is blank everytime I change the page. It first recieves the `value` of null and then the value it should have but the components doesn't rerender to show the actual `value`.